Added Custom Maya Scripts and Icons
Included here are a couple small mel scripts I wrote with the help of internet forums and searches. Also included are custom icons to correspond with those scripts as well as icons for scripts that can either be found for download online or are included as features within Maya. Links to external scripts are included in the README